+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 
 - PR #596 Add `tracking_memory_resource_adaptor` to help catch memory leaks
+- PR #608 Add stream wrapper type
 
 ## Improvements